Reduce the scope of the control-repo to mostly an example
Prior to this commit, the control-repo was an example of the
structure of a control repo but it also included puppet code to
help setup code manager and instructions to get that all setup
in a very specific way.

This was great for users that wanted to follow those instructions
exactly but wasn't great for people just looking for an example to
start from.

After this commit, the control-repo will be just an example once
again and a new repo somehwere else will pop up to provide the
explicit instructions on how to use the example with code manager.